Remove espionage (spy system) entirely
Step 2 of 4 in docs/SCO_SIMPLIFICATION_PLAN.md. Delete everything related
to spying, spy missions, fog of war, and counter-intel:

- Repositories/Spy/ (all 6 files), SpyMissionModule, SpyMission*,
  SpyAttemptLog, SpyResult, SpyCooldownException, FogOfWarRepository
- Spy* ViewModels in Shared, SpyController, Spies/SpyReports React pages
- PlayerState: SpyCooldowns, LastSpyResults, SpyAttemptLogs, SpyMissions
- GameNotificationType.SpyAttempted
- TechEffectType.CounterIntelDetection + .SpyEfficiency (dead code once
  spy repos are removed), plus the 3 counter-intel-* tech nodes in the
  SCO and Test game definition factories
- SpyCommand, SpyMissionCommand from Commands.cs
- spymission tick module registration
- BattleController / EnemyBaseViewModel SpyCostLabel path
- Frontend spy UI in SelectEnemy, EnemyBase, Help page
- All spy-related tests (SpyRepositoryTest, SpyMissionTest, SpyPlayersTest,
  SpyControllerTest, CounterIntelTest, FogOfWarRepositoryTest, three
  RaceMatchupBattleTest spy entries)

Built clean (0 warnings), 586 tests pass, ReactClient build clean.
Updates docs/SCO_GAME_LOGIC.md to match.
