+++ src/BrowserGameEngine.GameModel/PlayerStateImmutable.cs
@@ -19,14 +19,10 @@ namespace BrowserGameEngine.GameModel {
 		int UpgradeResearchTimer = 0,
 		UpgradeType UpgradeBeingResearched = UpgradeType.None,
 		IList<BuildQueueEntryImmutable>? BuildQueue = null,
-		IDictionary<string, DateTime>? SpyCooldowns = null,
 		IList<string>? UnlockedTechs = null,
 		string? TechBeingResearched = null,
 		int TechResearchTimer = 0,
-		IDictionary<string, SpyResult>? LastSpyResults = null,
-		IList<SpyAttemptLog>? SpyAttemptLogs = null,
 		IList<GameNotification>? Notifications = null,
-		IList<SpyMissionImmutable>? SpyMissions = null,
 		bool TutorialCompleted = false,
 		IList<ResourceSnapshot>? ResourceHistory = null,
 		IList<BattleReportImmutable>? BattleReports = null
