--- src/BrowserGameEngine.GameModel/SpyAttemptLog.cs
@@ -1,11 +1,0 @@
-using System;
-
-namespace BrowserGameEngine.GameModel {
-	public record SpyAttemptLog(
-		Guid Id,
-		PlayerId AttackerPlayerId,
-		string ActionType,
-		bool Detected,
-		DateTime Timestamp
-	);
-}
