+++ src/BrowserGameEngine.Shared/EnemyBaseViewModel.cs
@@ -8,6 +8,5 @@ namespace BrowserGameEngine.Shared {
 	public class EnemyBaseViewModel {
 		public required UnitsViewModel PlayerAttackingUnits { get; set; }
 		public required UnitsViewModel EnemyDefendingUnits { get; set; }
-		public string SpyCostLabel { get; set; } = "";
 	}
 }
