+++ src/BrowserGameEngine.StatefulGameServer/GameModelInternal/PlayerState.cs
@@ -22,14 +22,10 @@ namespace BrowserGameEngine.StatefulGameServer.GameModelInternal {
 		public int UpgradeResearchTimer { get; set; }
 		public UpgradeType UpgradeBeingResearched { get; set; }
 		public List<BuildQueueEntry> BuildQueue { get; set; } = new List<BuildQueueEntry>();
-		public Dictionary<string, DateTime> SpyCooldowns { get; set; } = new Dictionary<string, DateTime>();
 		public List<string> UnlockedTechs { get; set; } = new List<string>();
 		public string? TechBeingResearched { get; set; }
 		public int TechResearchTimer { get; set; }
-		public Dictionary<string, SpyResult> LastSpyResults { get; set; } = new Dictionary<string, SpyResult>();
-		public List<SpyAttemptLog> SpyAttemptLogs { get; set; } = new List<SpyAttemptLog>();
 		public List<GameNotification> Notifications { get; set; } = new List<GameNotification>();
-		public List<SpyMission> SpyMissions { get; set; } = new List<SpyMission>();
 		public bool TutorialCompleted { get; set; }
 		public List<ResourceSnapshot> ResourceHistory { get; set; } = new List<ResourceSnapshot>();
 		public List<BattleReport> BattleReports { get; set; } = new List<BattleReport>();
@@ -53,14 +49,10 @@ namespace BrowserGameEngine.StatefulGameServer.GameModelInternal {
 					UpgradeResearchTimer: playerState.UpgradeResearchTimer,
 					UpgradeBeingResearched: playerState.UpgradeBeingResearched,
 					BuildQueue: playerState.BuildQueue.Select(x => x.ToImmutable()).ToList(),
-					SpyCooldowns: new Dictionary<string, DateTime>(playerState.SpyCooldowns),
 					UnlockedTechs: new List<string>(playerState.UnlockedTechs),
 					TechBeingResearched: playerState.TechBeingResearched,
 					TechResearchTimer: playerState.TechResearchTimer,
-					LastSpyResults: new Dictionary<string, SpyResult>(playerState.LastSpyResults),
-					SpyAttemptLogs: new List<SpyAttemptLog>(playerState.SpyAttemptLogs),
 					Notifications: new List<GameNotification>(playerState.Notifications),
-					SpyMissions: playerState.SpyMissions.Select(x => x.ToImmutable()).ToList(),
 					TutorialCompleted: playerState.TutorialCompleted,
 					ResourceHistory: new List<ResourceSnapshot>(playerState.ResourceHistory),
 					BattleReports: playerState.BattleReports.Select(x => x.ToImmutable()).ToList()
@@ -85,25 +77,14 @@ namespace BrowserGameEngine.StatefulGameServer.GameModelInternal {
 				UpgradeBeingResearched = playerStateImmutable.UpgradeBeingResearched,
 				BuildQueue = (playerStateImmutable.BuildQueue ?? new List<BuildQueueEntryImmutable>())
 					.Select(x => x.ToMutable()).ToList(),
-				SpyCooldowns = playerStateImmutable.SpyCooldowns != null
-					? new Dictionary<string, DateTime>(playerStateImmutable.SpyCooldowns)
-					: new Dictionary<string, DateTime>(),
 				UnlockedTechs = playerStateImmutable.UnlockedTechs != null
 					? new List<string>(playerStateImmutable.UnlockedTechs)
 					: new List<string>(),
 				TechBeingResearched = playerStateImmutable.TechBeingResearched,
 				TechResearchTimer = playerStateImmutable.TechResearchTimer,
-				LastSpyResults = playerStateImmutable.LastSpyResults != null
-					? new Dictionary<string, SpyResult>(playerStateImmutable.LastSpyResults)
-					: new Dictionary<string, SpyResult>(),
-				SpyAttemptLogs = playerStateImmutable.SpyAttemptLogs != null
-					? new List<SpyAttemptLog>(playerStateImmutable.SpyAttemptLogs)
-					: new List<SpyAttemptLog>(),
 				Notifications = playerStateImmutable.Notifications != null
 					? new List<GameNotification>(playerStateImmutable.Notifications)
 					: new List<GameNotification>(),
-				SpyMissions = (playerStateImmutable.SpyMissions ?? new List<SpyMissionImmutable>())
-					.Select(x => x.ToMutable()).ToList(),
 				TutorialCompleted = playerStateImmutable.TutorialCompleted,
 				ResourceHistory = playerStateImmutable.ResourceHistory != null
 					? new List<ResourceSnapshot>(playerStateImmutable.ResourceHistory)
