--- src/BrowserGameEngine.StatefulGameServer/Repositories/Spy/SpyConstants.cs
@@ -1,8 +1,0 @@
-using System;
-
-namespace BrowserGameEngine.StatefulGameServer {
-	internal static class SpyConstants {
-		internal static readonly TimeSpan CooldownDuration = TimeSpan.FromMinutes(30);
-		internal const decimal SpyCostAmount = 50m;
-	}
-}
